% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exkludering.R
\name{landningssida_exkludera}
\alias{landningssida_exkludera}
\title{Lägg till appar/rapporter i exkluderingslistan}
\usage{
landningssida_exkludera(
  target = c("publik", "intern"),
  namn,
  andrad_av = NA_character_
)
}
\arguments{
\item{target}{\code{"publik"} eller \code{"intern"}.}

\item{namn}{Teckenvektor med mappnamn som ska exkluderas.}

\item{andrad_av}{Valfri identifierare för vem som gjorde ändringen (loggas i
DB, bara \code{"publik"}).}
}
\value{
Osynligt \code{TRUE}.
}
\description{
Lägg till appar/rapporter i exkluderingslistan
}
