% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ikoner.R
\name{landningssida_ikoner_lista}
\alias{landningssida_ikoner_lista}
\title{Lista manuella ikonkopplingar (inte standardgissningar)}
\usage{
landningssida_ikoner_lista(target = c("publik", "intern"))
}
\arguments{
\item{target}{\code{"publik"} eller \code{"intern"}.}
}
\value{
En \code{data.frame} med \code{namn} och \code{ikon}.
}
\description{
Lista manuella ikonkopplingar (inte standardgissningar)
}
