% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ikoner.R
\name{landningssida_ikoner_lista_tillgangliga}
\alias{landningssida_ikoner_lista_tillgangliga}
\title{Kurerad lista med relevanta Tabler-ikoner}
\usage{
landningssida_ikoner_lista_tillgangliga()
}
\value{
En \code{data.frame} med \code{ikon} och \code{beskrivning}.
}
\description{
Hela biblioteket (>4500 ikoner) går att bläddra på \url{https://tabler.io/icons};
\code{\link[=landningssida_ikoner_koppla]{landningssida_ikoner_koppla()}} accepterar vilken giltig \verb{ti-*}-klass som helst.
}
