% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ikoner.R
\name{landningssida_ikoner_ta_bort_koppling}
\alias{landningssida_ikoner_ta_bort_koppling}
\title{Ta bort en manuell ikonkoppling (återgår till standardgissningen)}
\usage{
landningssida_ikoner_ta_bort_koppling(target = c("publik", "intern"), namn)
}
\arguments{
\item{target}{\code{"publik"} eller \code{"intern"}.}

\item{namn}{Mappnamnet vars koppling ska tas bort.}
}
\value{
Osynligt \code{TRUE}.
}
\description{
Ta bort en manuell ikonkoppling (återgår till standardgissningen)
}
