% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exkludering.R
\name{landningssida_inkludera}
\alias{landningssida_inkludera}
\title{Ta bort appar/rapporter från exkluderingslistan}
\usage{
landningssida_inkludera(target = c("publik", "intern"), namn)
}
\arguments{
\item{target}{\code{"publik"} eller \code{"intern"}.}

\item{namn}{Teckenvektor med mappnamn som ska exkluderas.}
}
\value{
Osynligt \code{TRUE}.
}
\description{
Ta bort appar/rapporter från exkluderingslistan
}
