% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exkludering.R
\name{landningssida_lista_exkluderade}
\alias{landningssida_lista_exkluderade}
\title{Lista exkluderade appar/rapporter}
\usage{
landningssida_lista_exkluderade(target = c("publik", "intern"))
}
\arguments{
\item{target}{\code{"publik"} eller \code{"intern"}.}
}
\value{
Teckenvektor med mappnamn (osynligt tom vektor om listan är tom).
}
\description{
Lista exkluderade appar/rapporter
}
