% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nedladdning.R
\name{landningssida_nedladdning_lista}
\alias{landningssida_nedladdning_lista}
\title{Aggregerad översikt över nedladdningsfiler (en rad per app)}
\usage{
landningssida_nedladdning_lista(target = c("publik", "intern"))
}
\arguments{
\item{target}{\code{"publik"} eller \code{"intern"}.}
}
\value{
En \code{data.frame} med \code{app}, \code{antal_filer}, \code{senaste_andring},
\code{total_storlek}.
}
\description{
Aggregerad översikt över nedladdningsfiler (en rad per app)
}
