% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nedladdning.R
\name{landningssida_nedladdning_lista_filer}
\alias{landningssida_nedladdning_lista_filer}
\title{Detaljlista över nedladdningsfiler för en app}
\usage{
landningssida_nedladdning_lista_filer(target = c("publik", "intern"), app)
}
\arguments{
\item{target}{\code{"publik"} eller \code{"intern"}.}

\item{app}{Appnamnet.}
}
\value{
En \code{data.frame} med \code{filnamn}, \code{storlek_bytes}, \code{senast_andrad}.
}
\description{
Detaljlista över nedladdningsfiler för en app
}
