% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nedladdning.R
\name{landningssida_nedladdning_ta_bort_fil}
\alias{landningssida_nedladdning_ta_bort_fil}
\title{Ta bort en enskild nedladdningsfil (och mappen om den blir tom)}
\usage{
landningssida_nedladdning_ta_bort_fil(
  target = c("publik", "intern"),
  app,
  filnamn,
  andrad_av = NA_character_
)
}
\arguments{
\item{target}{\code{"publik"} eller \code{"intern"}.}

\item{app}{Appnamnet.}

\item{filnamn}{Filnamnet.}

\item{andrad_av}{Valfri identifierare (bara \code{"publik"}).}
}
\value{
Osynligt \code{TRUE}.
}
\description{
Ta bort en enskild nedladdningsfil (och mappen om den blir tom)
}
