% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nedladdning.R
\name{landningssida_nedladdning_ta_bort_mapp}
\alias{landningssida_nedladdning_ta_bort_mapp}
\title{Ta bort hela nedladdningsmappen för en app}
\usage{
landningssida_nedladdning_ta_bort_mapp(
  target = c("publik", "intern"),
  app,
  andrad_av = NA_character_
)
}
\arguments{
\item{target}{\code{"publik"} eller \code{"intern"}.}

\item{app}{Appnamnet.}

\item{andrad_av}{Valfri identifierare (bara \code{"publik"}).}
}
\value{
Osynligt \code{TRUE}.
}
\description{
Ta bort hela nedladdningsmappen för en app
}
