% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdadminportal-package.R
\docType{package}
\name{rdadminportal-package}
\alias{rdadminportal}
\alias{rdadminportal-package}
\title{rdadminportal: Adminportal och Serverdrift för Region Dalarnas Shiny-landningssidor}
\description{
Funktioner för adminportalens Exkludering-, Ikoner-, Appar- och Nedladdning-flikar samt för de serverskript som synkar Shiny- landningssidorna. Hanterar VILKA appar och rapporter som visas, deras ikoner och exkluderingsstatus, mot den delade adminshiny-databasen ("publik") eller lokala sudo-anrop ("intern"). Innehåller även kor_cron_jobb() som exekverar ett schemalagt datauttag. Utbrutet ur func_landningssida_adminportal.R och func_kor_cron_jobb.R i Region-Dalarna/funktioner.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Region-Dalarna/rdpaket}
  \item Report bugs at \url{https://github.com/Region-Dalarna/rdpaket/issues}
}

}
\author{
\strong{Maintainer}: Peter Möller \email{larspetermoller@gmail.com}

Authors:
\itemize{
  \item Peter Möller \email{larspetermoller@gmail.com}
}

Other contributors:
\itemize{
  \item Region Dalarna [copyright holder]
}

}
\keyword{internal}
